pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentmodelctl"
version = "0.2.0"
description = "Fleet management for AI agents — know what breaks before you switch models"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "agentmodelctl contributors" }]
keywords = ["ai", "agents", "llm", "model-management", "evaluation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    "typer>=0.9",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "litellm>=1.0",
    "rich>=13.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.4",
]

[project.scripts]
agentmodelctl = "agentmodelctl.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/agentmodelctl"]

[tool.ruff]
src = ["src"]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
